+++ src/main/java/net/minestom/server/adventure/provider/MinestomComponentLoggerProvider.java
@@ -2,24 +2,13 @@ package net.minestom.server.adventure.provider;
 
 import net.kyori.adventure.text.logger.slf4j.ComponentLogger;
 import net.kyori.adventure.text.logger.slf4j.ComponentLoggerProvider;
-import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
-import net.minestom.server.terminal.TerminalColorConverter;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.LoggerFactory;
 
 @SuppressWarnings("UnstableApiUsage") // we are permitted to provide this
 public class MinestomComponentLoggerProvider implements ComponentLoggerProvider {
-    private static final LegacyComponentSerializer serializer = LegacyComponentSerializer.builder()
-            .character(TerminalColorConverter.COLOR_CHAR)
-            .flattener(MinestomFlattenerProvider.INSTANCE)
-            .hexColors()
-            .build();
-
     @Override
     public @NotNull ComponentLogger logger(@NotNull LoggerHelper helper, @NotNull String name) {
-        return helper.delegating(LoggerFactory.getLogger(name), component -> {
-            String text = serializer.serialize(component);
-            return TerminalColorConverter.format(text);
-        });
+        return helper.delegating(LoggerFactory.getLogger(name), TerminalColorConverter::format);
     }
 }
