+++ src/main/java/net/minestom/server/adventure/provider/TerminalColorConverter.java
@@ -1,6 +1,8 @@
-package net.minestom.server.terminal;
+package net.minestom.server.adventure.provider;
 
+import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
+import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 import net.minestom.server.utils.PropertyUtils;
 
 import java.util.regex.Matcher;
@@ -14,6 +16,11 @@ import java.util.regex.Pattern;
  */
 public final class TerminalColorConverter {
     public static final char COLOR_CHAR = '§';
+    private static final LegacyComponentSerializer SERIALIZER = LegacyComponentSerializer.builder()
+            .character(TerminalColorConverter.COLOR_CHAR)
+            .flattener(MinestomFlattenerProvider.INSTANCE)
+            .hexColors()
+            .build();
     private static final boolean SUPPORT_HEX_COLOR = PropertyUtils.getBoolean("minestom.terminal.support-hex-color", true);
     private static final String RGB_ANSI = "\u001B[38;2;%d;%d;%dm";
     private static final String ANSI_RESET = "\u001B[m";
@@ -44,6 +51,9 @@ public final class TerminalColorConverter {
     };
     private static final Pattern RGB_PATTERN = Pattern.compile(COLOR_CHAR + "#([\\da-fA-F]{6})");
     private static final Pattern NAMED_PATTERN = Pattern.compile(COLOR_CHAR + "([\\da-fk-orA-FK-OR])");
+
+    private TerminalColorConverter() {
+    }
 
     private static String getAnsiColor(NamedTextColor color, String fallback) {
         return SUPPORT_HEX_COLOR ? getAnsiColorFromHexColor(color.value()) : fallback;
@@ -82,5 +92,15 @@ public final class TerminalColorConverter {
 
         builder.append(ANSI_RESET);
         return builder.toString();
+    }
+
+    /**
+     * Format the component to an ansi-colored string.
+     *
+     * @param component the component to format
+     * @return the formatted string
+     */
+    public static String format(Component component) {
+        return format(SERIALIZER.serialize(component));
     }
 }
